fix: tighten router wildcard and cross-platform storage/cluster keys
Require trailing # in topic patterns, format cluster node ids as
ullong, write storage lengths as uint64, and reset kqueue filters on
mod so WRITE does not stick after drain.

+++ src/cluster/t_cluster.c
@@ -12,7 +12,7 @@ struct t_cluster {
 };
 
 static void make_key(char *buf, size_t sz, uint64_t id) {
-    snprintf(buf, sz, "%lu", (unsigned long)id);
+    snprintf(buf, sz, "%llu", (unsigned long long)id);
 }
 
 t_cluster *t_cluster_create(uint64_t self_id) {
